different types of charts in different moduls from now on
