+++ vis.py
@@ -8,8 +8,30 @@ from processing import renishaw
 from processing import utils
 from processing import wasatch
 from processing import witec
-from visualisation import custom_plot
 from visualisation import draw
+from visualisation import grouped_spectra as gs
+from visualisation import mean_spectra as ms
+from visualisation import p3d_spectra as p3d
+from visualisation import sigle_spectra as ss
+from visualisation import visualisation_options as vo
+
+SINGLE = 'Single spectra'
+MS = "Mean spectrum"
+GS = "Grouped spectra"
+P3D = "Plot 3D"
+RAW = "Raw Data"
+OPT = "Optimised Data"
+NORM = "Normalized"
+
+UplSpec = 'Upload "*.txt" spectra'
+BWTEK = 'BWTEK'
+RENI = 'Renishaw'
+WITEC = 'WITec Alpha300 R+'
+WASATCH = 'Wasatch System'
+TELEDYNE = 'Teledyne Princeton Instruments'
+SEPARATORS = {'comma': ',', 'dot': '.', 'tab': '\t', 'space': ' '}
+
+TEST = 'Testing bwtek'
 
 st.set_page_config(
     page_title="SERSitive.eu",
@@ -39,19 +61,6 @@ st.sidebar.markdown(html_code, unsafe_allow_html=True)
 st.sidebar.markdown('')
 st.sidebar.markdown('')
 
-SINGLE = 'Single spectra'
-MS = 'Mean spectrum'
-GS = 'Grouped spectra'
-P3D = 'Plot 3D'
-UplSpec = 'Upload "*.txt" spectra'
-BWTEK = 'BWTEK'
-RENI = 'Renishaw'
-WITEC = 'WITec Alpha300 R+'
-WASATCH = 'Wasatch System'
-TELEDYNE = 'Teledyne Princeton Instruments'
-SEPARATORS = {'comma': ',', 'dot': '.', 'tab': '\t', 'space': ' '}
-
-TEST = 'Testing bwtek'
 
 st.sidebar.write('#### Choose spectra type', unsafe_allow_html=True)
 spectrometer = st.sidebar.radio(
@@ -77,53 +86,53 @@ if files:
         plots_color = draw.choosing_colorway()
         template = draw.choose_template()
 
+    # BWTek raw spectra
     if spectrometer == BWTEK:
         temp_data_df, temp_meta_df = bwtek.read_bwtek(files)
         df = utils.group_dfs(temp_data_df)
-        custom_plot.bwtek_vis_options(df, plots_color, template)
 
     # Renishaw raw spectra
     elif spectrometer == RENI:
-    
         reni_data = renishaw.read_renishaw(files, SEPARATORS['space'])
-
         df = pd.concat([reni_data[data_df] for data_df in reni_data], axis=1)
-
         df.dropna(inplace=True, how='any', axis=0)
-
-        display_opt = custom_plot.vis_options()
-        custom_plot.show_plot(df, plots_color, template, display_opt=display_opt)
 
     # WITec raw spectra
     elif spectrometer == WITEC:
         witec_data = witec.read_witec(files, SEPARATORS['comma'])
-
         df = pd.concat([witec_data[data_df] for data_df in witec_data], axis=1)
-
-        display_opt = custom_plot.vis_options()
-        custom_plot.show_plot(df, plots_color, template, display_opt=display_opt)
 
     elif spectrometer == WASATCH:
         # Read data and prepare it for plot
         data = wasatch.read_wasatch(files, SEPARATORS['comma'])
 
-        # Show possible options for visualisation - single/grouped spectra
-        display_opt = custom_plot.vis_options()
-
-        # Plot spectra
-        custom_plot.show_plot(data, plots_color, template, display_opt=display_opt)
-
     # Renishaw raw spectra
     elif spectrometer == TELEDYNE:
-    
         reni_data = renishaw.read_renishaw(files, SEPARATORS['comma'])
-    
         df = pd.concat([reni_data[data_df] for data_df in reni_data], axis=1)
-    
         df.dropna(inplace=True, how='any', axis=0)
-    
-        display_opt = custom_plot.visualisation_options()
-        custom_plot.show_plot(df, plots_color, template, display_opt=display_opt)
+
+    # lets you select chart type
+    chart_type = vo.vis_options(spectrometer)
+
+    # lets you select data conversion type
+    st.sidebar.write('#### How would you like to convert the data?', unsafe_allow_html=True)
+    spectra_conversion_type = st.sidebar.radio(
+        "",
+        (RAW, OPT, NORM), key=f'raw'
+    )
+
+    params = plots_color, template, chart_type, spectra_conversion_type
+
+    # All possible types of charts
+    data_vis_option = {SINGLE: ss.show_single_plots,
+                       MS: ms.show_mean_plot,
+                       GS: gs.show_grouped_plot,
+                       P3D: p3d.show_3d_plots}
+
+    # run specified type of chart with chosen parameters
+    data_vis_option[chart_type](df, params)
+
 
 else:
     st.markdown(f'''
