+++ visualisation/__init__.py
@@ -1,3 +1,6 @@
-from . import custom_plot
 from . import draw
+from . import grouped_spectra
+from . import mean_spectra
+from . import p3d_spectra
+from . import sigle_spectra
 from . import visualisation_options
